pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "map2-patch-discovery"
version = "0.1.0"
description = "MAP2-ROI-driven patch extraction and discovery scaffolding for ICC puncta analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "PyYAML>=6.0",
  "tifffile>=2024.2.12",
  "zarr>=2.18",
  "scipy>=1.11",
  "scikit-learn>=1.5",
  "matplotlib>=3.8",
]

[project.scripts]
map2-patch-extract = "map2_patch_discovery.cli:main"
map2-latent-report = "map2_patch_discovery.report_cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
